fix: refactor MultiSigRequest
