+++ src/ui/pages/IncomingRequest/components/MultiSigRequest.scss
