+++ src/ui/pages/IncomingRequest/components/RequestComponent.test.tsx
@@ -116,7 +116,7 @@ describe("Multi-Sig request", () => {
     );
 
     expect(
-      getByText(EN_TRANSLATIONS.request.multisig.stageone.title)
+      getByText(EN_TRANSLATIONS.request.multisig.title)
     ).toBeInTheDocument();
   });
 
@@ -162,7 +162,7 @@ describe("Multi-Sig request", () => {
     );
 
     expect(
-      getByText(EN_TRANSLATIONS.request.multisig.stageone.title)
+      getByText(EN_TRANSLATIONS.request.multisig.title)
     ).toBeInTheDocument();
 
     expect(queryByTestId("multisig-connection-fallback-logo")).toBeVisible();
