+++ src/ui/pages/IncomingRequest/components/RequestComponent.tsx
@@ -1,9 +1,7 @@
-import { useState } from "react";
 import { IncomingRequestType } from "../../../../store/reducers/stateCache/stateCache.types";
 import { RequestProps } from "../IncomingRequest.types";
 import { CredentialRequest } from "./CredentialRequest";
-import { MultiSigRequestStageOne } from "./MultiSigRequestStageOne";
-import { MultiSigRequestStageTwo } from "./MultiSigRequestStageTwo";
+import { MultiSigRequest } from "./MultiSigRequest";
 
 const RequestComponent = ({
   pageId,
@@ -17,7 +15,6 @@ const RequestComponent = ({
   handleIgnore,
   incomingRequestType,
 }: RequestProps) => {
-  const [requestStage, setRequestStage] = useState(0);
   switch (incomingRequestType) {
   case IncomingRequestType.CREDENTIAL_OFFER_RECEIVED:
     return (
@@ -32,8 +29,10 @@ const RequestComponent = ({
       />
     );
   case IncomingRequestType.MULTI_SIG_REQUEST_INCOMING:
-    return requestStage === 0 ? (
-      <MultiSigRequestStageOne
+    return (
+      <MultiSigRequest
+        blur={blur}
+        setBlur={setBlur}
         pageId={pageId}
         activeStatus={activeStatus}
         requestData={requestData}
@@ -41,19 +40,6 @@ const RequestComponent = ({
         handleAccept={handleAccept}
         handleCancel={handleCancel}
         handleIgnore={handleIgnore}
-        setRequestStage={setRequestStage}
-      />
-    ) : (
-      <MultiSigRequestStageTwo
-        pageId={pageId}
-        activeStatus={activeStatus}
-        blur={blur}
-        setBlur={setBlur}
-        requestData={requestData}
-        initiateAnimation={initiateAnimation}
-        handleAccept={handleAccept}
-        handleCancel={handleCancel}
-        setRequestStage={setRequestStage}
       />
     );
   default:
